#33, #32: World and entity physics
Squashed commit of the following:

commit a25183c7e27efe7d21f4817a4c18fa17837ab986
Author: Dom Williams <me@domwillia.ms>
Date:   Fri Nov 22 08:27:49 2019 +0000

    #33: Tweak tweaker values and entity spawn location

commit 2152d992b7989638e271affb814bf5b2db7a1150
Author: Dom Williams <me@domwillia.ms>
Date:   Sun Nov 17 16:36:14 2019 +0000

    #33: Remove empty test

commit 4a61f3064f73265a375adf9d8164ffdc6f67a3a3
Author: Dom Williams <me@domwillia.ms>
Date:   Sun Nov 17 16:30:48 2019 +0000

    #33: Add toggle for physics debug rendering

commit f40d62d2fc27673886427f0e3a25c601de4a424e
Author: Dom Williams <me@domwillia.ms>
Date:   Sun Nov 17 16:22:52 2019 +0000

    #33, #32: Fix greedy meshing

commit 9087900347641c9cab3f59487d4def62f3679b0b
Author: Dom Williams <me@domwillia.ms>
Date:   Sun Nov 17 16:22:30 2019 +0000

    #33, #32: Fix incorrect slab slice resolution

commit bce99fcc8bdce98c95c94f438d5606b103fae3a5
Author: Dom Williams <me@domwillia.ms>
Date:   Sun Nov 17 13:14:29 2019 +0000

    #33, #32: Fix slab mesh position

commit 0e89e682e15e2321e227bea9496007834b752e81
Author: Dom Williams <me@domwillia.ms>
Date:   Sun Nov 17 12:45:16 2019 +0000

    #33: Add crazy ffi adventure for bullet debug rendering

commit aee82bd5875a6c4d29748a87f253dff94fbb3d68
Author: Dom Williams <me@domwillia.ms>
Date:   Sun Nov 17 09:31:33 2019 +0000

    #33: Add not fully working entity colliders

commit 02eefbe23536b59c9f3c27c87aafa338b012a810
Author: Dom Williams <me@domwillia.ms>
Date:   Sun Nov 17 08:35:14 2019 +0000

    #33, #32: Add slab collision mesh creation

commit 9e87d17da1c10d8bf3483ddf474f43e403a3dc48
Author: Dom Williams <me@domwillia.ms>
Date:   Sat Nov 16 12:23:12 2019 +0000

    #33: Extract ffi to bulletc-sys crate

commit 3056eaa15e5c7f8455ffb065fc7ee0211a0f3c02
Author: Dom Williams <me@domwillia.ms>
Date:   Sat Nov 16 11:53:01 2019 +0000

    #33: Add simple bullet world creation through ffi

commit 415a160b672cf02a51d93e3c32191722bb4b101a
Author: Dom Williams <me@domwillia.ms>
Date:   Sat Nov 16 11:12:24 2019 +0000

    #33: Add bindgen for bulletc bindings

commit bae448b229787d48243eeb682cd5f9fb065845b6
Author: Dom Williams <me@domwillia.ms>
Date:   Sat Nov 16 10:47:48 2019 +0000

    #33: Add empty bullet ffi lib

commit 976cdf7fe42855c592d8c9b08dc3f064113ba047
Author: Dom Williams <me@domwillia.ms>
Date:   Fri Nov 15 14:59:56 2019 +0000

    #33: Extract coordinate module to unit crate

commit cf29397dd1cc64d36e3f092bc9935e56ddf10f5b
Author: Dom Williams <me@domwillia.ms>
Date:   Fri Nov 15 14:45:07 2019 +0000

    #33: Purge the horrendous nphysics

commit c1c1357c16a43e7338df050196834a767c96fe4d
Author: Dom Williams <me@domwillia.ms>
Date:   Fri Nov 15 14:41:35 2019 +0000

    #33: Add rustfmt config

commit 19d113724d6d5deb4627fe08bead6758285cb359
Author: Dom Williams <me@domwillia.ms>
Date:   Fri Nov 15 14:17:11 2019 +0000

    #33: Add world collision mesh generation

commit 155c17cbd9dfa882bece1c6fcfa92b64ef88fbae
Author: Dom Williams <me@domwillia.ms>
Date:   Sun Nov 10 10:10:58 2019 +0000

    #33: Add nphysics

commit c28bffe4cbdd2646cb6c3579d720cf4c16c5ef1a
Author: Dom Williams <me@domwillia.ms>
Date:   Sat Nov 9 16:56:41 2019 +0000

    #33: Fix confusion between world and view space

commit 5b42172496d8be0c2acd5f0fa8b3550c733871cf
Author: Dom Williams <me@domwillia.ms>
Date:   Sat Nov 9 16:55:52 2019 +0000

    #33: Add wireframe toggle

commit 00f80a872586e26bbcbd824e2a91786f25017d88
Author: Dom Williams <me@domwillia.ms>
Date:   Sat Nov 9 14:36:36 2019 +0000

    #33: Add rudimentary physics system

commit 535e3f34d0e368630ebe3ab1fc2c3837e3006c0f
Author: Dom Williams <me@domwillia.ms>
Date:   Sat Nov 9 12:03:57 2019 +0000

    #33: Render entities as scaled cuboids

commit 6351f77f9291a0b7848a4977f333ff0707283442
Author: Dom Williams <me@domwillia.ms>
Date:   Sat Nov 9 11:55:52 2019 +0000

    #33: Render entities as cubes

+++ simulation/src/steer/behaviour.rs
@@ -8,9 +8,9 @@ use world::WorldPoint;
 impl From<Position> for Vector3<f32> {
     fn from(pos: Position) -> Self {
         Self {
-            x: pos.x,
-            y: pos.y,
-            z: pos.z as f32,
+            x: pos.x(),
+            y: pos.y(),
+            z: pos.z(),
         }
     }
 }
